+++ src/Authentications/RefreshToken.cs
@@ -1,4 +1,6 @@
-﻿namespace Basecamp3Api;
+﻿using System.Collections.Specialized;
+
+namespace Basecamp3Api;
 
 public partial class BasecampApiClient
 {
@@ -8,37 +10,37 @@ public partial class BasecampApiClient
     /// <returns></returns>
     /// <exception cref="InvalidOperationException"></exception>
     /// <exception cref="Exception"></exception>
-    public async Task<Token> RefreshTokenAsync(CancellationToken cancellationToken = default)
+    public async Task<(Token? Token, Error? Error)> RefreshTokenAsync(CancellationToken cancellationToken = default)
     {
         if (!TokenHasBeenSet)
             throw new InvalidOperationException("Token has not been set");
 
         //POST https://launchpad.37signals.com/authorization/token?type=refresh&refresh_token=your-current-refresh-token&client_id=your-client-id&redirect_uri=your-redirect-uri&client_secret=your-client-secret
 
-        var tokenUriBuilder = new UriBuilder(AuthTokenUrl);
-        var query = HttpUtility.ParseQueryString(tokenUriBuilder.Query);
-        query["type"] = "refresh";
-        query["client_id"] = _setting.ClientId;
-        query["redirect_uri"] = _setting.RedirectUrl!.OriginalString;
-        query["client_secret"] = _setting.ClientSecret;
-        query["refresh_token"] = RefreshToken;
-        tokenUriBuilder.Query = query.ToString();
+        var uriBuilder = new UriBuilder(AuthTokenUrl);
+        var nvc = new NameValueCollection(5)
+        {
+            ["type"] = "refresh",
+            ["client_id"] = _setting.ClientId,
+            ["redirect_uri"] = _setting.RedirectUrl!.OriginalString,
+            ["client_secret"] = _setting.ClientSecret,
+            ["refresh_token"] = RefreshToken
+        };
+        uriBuilder.AddQueryParams(ConstructQueryString(nvc));
 
-        var request = new HttpRequestMessage(HttpMethod.Post, tokenUriBuilder.ToString());
+        var request = CreateRequestMessage(HttpMethod.Post, uriBuilder.Uri, null);
 
-        var response = await _httpClient.SendAsync(request, cancellationToken);
+        var response = await SendMessageAsync(request, HttpStatusCode.OK, cancellationToken);
 
-        var content = await response.Content.ReadAsStringAsync(cancellationToken);
+        if (response.Error != null)
+            return (null, Error: response.Error);
 
-        if (response.StatusCode != HttpStatusCode.OK)
-            throw new Exception("Result not OK when Refresh Token", new Exception($"With message : {content}"));
-
-        var result = JsonSerializer.Deserialize<Token>(content)!;
+        var result = JsonSerializer.Deserialize<Token>(response.Response!.Value.ResultJsonInString)!;
 
         //re-update current access token
         AccessToken = result.AccessToken;
         ExpiresIn = result.ExpiresIn;
 
-        return result;
+        return (result, null);
     }
 }