+++ src/BasecampApiSetting.cs
@@ -4,5 +4,6 @@ public class BasecampApiSetting
 {
     public string? ClientId { get; set; }
     public string? ClientSecret { get; set; }
+    public string? AppName { get; set; }
     public Uri? RedirectUrl { get; set; }
 }