+++ src/Exceptions/InsufficientStorageException.cs
@@ -1,9 +1,0 @@
-﻿namespace Basecamp3Api.Exceptions;
-
-public class InsufficientStorageException : Exception
-{
-    public InsufficientStorageException() : base(
-        "The project limit for this account has been reached")
-    {
-    }
-}