+++ src/Extensions.cs
@@ -1,0 +1,10 @@
+﻿namespace Basecamp3Api;
+
+internal static class Extensions
+{
+    internal static UriBuilder AddQueryParams(this UriBuilder uriBuilder, string s)
+    {
+        uriBuilder.Query = s;
+        return uriBuilder;
+    }
+}