+++ src/Usings.cs
@@ -1,5 +1,4 @@
-﻿global using Basecamp3Api.Exceptions;
-global using Basecamp3Api.Models;
+﻿global using Basecamp3Api.Models;
 global using System.Net.Http.Headers;
 global using System.Text.Json;
 global using System.Text.Json.Serialization;