+++ tests/Projects/GetAllProjectTests.cs
@@ -1,4 +1,6 @@
-﻿namespace Basecamp3Api.Tests.Projects;
+﻿using Basecamp3Api.Models;
+
+namespace Basecamp3Api.Tests.Projects;
 
 [Collection(nameof(BaseFixture))]
 public class GetAllProjectTests
@@ -14,16 +16,16 @@ public class GetAllProjectTests
     public async Task GetAllProject_Should_Do_As_Expected()
     {
         var auth = await _baseFixture.Client.GetAuthorizationAsync();
-        auth.ShouldNotBeNull();
+        auth.Auth.ShouldNotBeNull();
 
         var page = 1;
-        var projects = await _baseFixture.Client.GetAllProjectAsync(auth.Accounts!.First().Id, 1);
-        projects.Results.Count.ShouldBePositive();
+        var projects = await _baseFixture.Client.GetAllProjectAsync(auth.Auth.Accounts!.First().Id, 1);
+        projects.List!.Results.Count.ShouldBePositive();
 
-        if (projects.HasNextPage)
+        if (projects.List.HasNextPage)
         {
-            projects = await _baseFixture.Client.GetAllProjectAsync(auth.Accounts!.First().Id, ++page);
-            projects.Results.Count.ShouldBePositive();
+            projects = await _baseFixture.Client.GetAllProjectAsync(auth.Auth.Accounts!.First().Id, ++page);
+            projects.List!.Results.Count.ShouldBePositive();
         }
     }
 }