+++ CHANGELOG.md
@@ -2,11 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.1.0-alpha
+
+- Remove exceptions, use tuple instead
+- Adding Error Struct and value as results
+- Bugfix and improvement
+- Add new setting, Application Name
+
 ## 0.0.3-alpha
 
 - Add project API
-  - Update Project
-  - Trash Project
+    - Update Project
+    - Trash Project
 
 ## 0.0.2-alpha